fix : added global exception handler

+++ src/main/java/com/hashedin/user_service/controller/UserController.java
@@ -1,10 +1,13 @@
 package com.hashedin.user_service.controller;
 
+import com.hashedin.user_service.exception.handler.RoleNotAllowed;
+import com.hashedin.user_service.exception.handler.UserNotAuthorised;
 import com.hashedin.user_service.model.RegisterUser;
 import com.hashedin.user_service.model.User;
 import com.hashedin.user_service.service.UserService;
-import io.swagger.v3.oas.annotations.parameters.RequestBody;
+import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.apache.coyote.BadRequestException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -18,6 +21,7 @@ import java.util.List;
 @RequestMapping("/users")
 @RestController
 @SecurityRequirement(name = "bearerAuth")
+@Tag(name = "User Control")
 public class UserController {
 
     private final UserService userService;
@@ -28,31 +32,52 @@ public class UserController {
         this.userService = userService;
     }
 
-    @GetMapping("/me")
-    public ResponseEntity<User> authenticatedUser() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        User currentUser = (User) authentication.getPrincipal();
+    @GetMapping("/{id}")
+    @Operation(summary = " Current user details", description = " To see details of the logged in user")
+    public ResponseEntity<User> authenticatedUser(@PathVariable("id") int id) {
+        User currentUser = userService.findUserById(id);
         return ResponseEntity.ok(currentUser);
     }
 
     @DeleteMapping("/delete")
-    public ResponseEntity<String> deleteUserByEmail(@RequestParam String email){
+    @Operation(summary = " Delete Users", description = " To delete user by email ")
+    public ResponseEntity<String> deleteUserByEmail(@RequestParam String email)  throws BadRequestException {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+        if(!currentUser.getEmail().equals(email)){
+            throw new BadRequestException(" Can't delete other user account ");
+        }
         userService.deleteUserByEmail(email);
         return ResponseEntity.ok(" User with email " + email + " deleted successfully");
     }
 
-    @PutMapping("/update/{email}")
-    public ResponseEntity<User> updateUserByEmail(@PathVariable("email") String email, @RequestBody RegisterUser user) throws BadRequestException {
-        System.out.println(user.getName());
-        if(!roles.contains(user.getRole())){
-            throw new BadRequestException(" Role can only be in { ADMIN, CUSTOMER, RESTAURANT }");
+    @PutMapping("/update")
+    @Operation(summary = " Update user details ", description = " To update the users details ")
+    public ResponseEntity<User> updateUserByEmail(@RequestParam String email, @RequestBody RegisterUser registerUser) {
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+
+        if(!currentUser.getEmail().equals(email)){
+            throw new UserNotAuthorised(" Can't update other user account details");
         }
-            User updatedUser = userService.updateUserByEmail(email, user);
+
+        if(!roles.contains(registerUser.getRole())){
+            throw new RoleNotAllowed(" Role can only be in { ADMIN, CUSTOMER, RESTAURANT }");
+        }
+            User updatedUser = userService.updateUserByEmail(email, registerUser);
             return ResponseEntity.ok(updatedUser);
     }
 
     @PatchMapping("/update-password")
-    public ResponseEntity<String> updatePasswordByEmail(@RequestParam String email, @RequestParam String newPassword){
+    @Operation(summary = " Reset Password ", description = " If user forgets password.. use option to reset it.")
+    public ResponseEntity<String> updatePasswordByEmail(@RequestParam String email, @RequestParam String newPassword) throws BadRequestException{
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+
+        if(!currentUser.getEmail().equals(email)){
+            throw new UserNotAuthorised(" Can't delete other user account ");
+        }
         return ResponseEntity.ok(userService.updatePassword(email, newPassword));
     }
